% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnnDeltaVariance.R
\name{mnnDeltaVariance}
\alias{mnnDeltaVariance}
\title{Computes the variance of the paired MNN deltas}
\usage{
mnnDeltaVariance(
  ...,
  pairs,
  subset.row = NULL,
  BPPARAM = SerialParam(),
  trend.args = list()
)
}
\arguments{
\item{...}{One or more log-expression matrices where genes correspond to rows and cells correspond to columns.
Alternatively, one or more \linkS4class{SingleCellExperiment} objects can be supplied containing a log-expression matrix in the \code{assay.type} assay.
Each object should contain the same number of rows, corresponding to the same genes in the same order.
Objects of different types can be mixed together.

If multiple objects are supplied, each object is assumed to contain all and only cells from a single batch.
If a single object is supplied, it is assumed to contain cells from all batches, so \code{batch} should also be specified.

Alternatively, one or more lists of matrices or SingleCellExperiments can be provided;
this is flattened as if the objects inside each list were passed directly to \code{...}.}

\item{pairs}{A \linkS4class{DataFrame} or list of \linkS4class{DataFrame}s containing MNN pairing information.
Each row of each DataFrame specifies an MNN pair; each DataFrame should have two columns containing the column indices of paired cells.
This is typically produced by \code{\link{fastMNN}}, see that documentation for more information.}

\item{subset.row}{A vector specifying which features to use for correction.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether the PCA and nearest-neighbor searches should be parallelized.}

\item{trend.args}{Named list of further arguments to pass to \code{fitTrendVar} from the \pkg{scran} package.}
}
\value{
A numeric vector of variances for each row in the input objects (or as specified by \code{subset.row}).
}
\description{
Computes the variance of the paired MNN deltas
}
\details{
The \dQuote{MNN delta} is defined as the difference in the \emph{uncorrected} expression values of MNN-paired cells.
We compute the variance of these values across all pairs for each gene; genes with highly variable deltas are strongly affected by the correction beyond a simple shift.
By looking through the top genes, we may conclude that the correction did something unusual if, e.g., we find important markers in the top set.
Conversely, if key genes do not have unusually high variances in their delta, we may gain some confidence in the reliability of the correction -
at least with respect to the behavior of those genes.

We compute the variance of the deltas rather than using their magnitude, 
as a shift in expression space is not particularly concerning (and is, in fact, par for the course for batch correction).
Rather, we are interested in identifying non-linear changes that might be indicative of correction errors, e.g., inappropriate merging of two different subpopulations.
This would manifest as high variances for the relevant marker genes.
Of course, whether or not this is actually an error can be a subjective decision - loss of some biological heterogeneity is often an acceptable cost for flexible correction.
}
\examples{
means <- 2^rexp(200) * 10
B1 <- matrix(rpois(10000, means), ncol=50) # Batch 1 
B2 <- matrix(rpois(10000, means), ncol=50) # Batch 2

# Spiking in some population structure. Each batch is split into two halves 
# corresponding to 2 subpopulations. These populations match up across batches
# but the group in batch 1 also upregulates the first gene.
B1[1:2,1:25] <- B1[1:2,1:25] * 50 
B2[2,1:25] <- B2[2,1:25] * 50 

# Quick log-transformation and correction. We'll omit the multiBatchNorm as the
# sequencing depth is the same across batches anyway.
library(scuttle)
B1 <- normalizeCounts(B1)
B2 <- normalizeCounts(B2)
out <- fastMNN(B1, B2)

# First gene should have high variance of deltas, as the upregulation in B1 is 
# removed to enable the alignment of subpopulations across batches.
mnnDeltaVariance(B1, B2, pairs=metadata(out)$merge.info$pairs[[1]])

}
\author{
Aaron Lun
}
